+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Yii Framework 2 bootstrap extension Change Log
 2.0.10 under development
 ------------------------
 
-- no changes in this release.
+- Bug #253: Missing initial class `collapsed` of an `a` element in heading in `yii\bootstrap\Collapse` widget.
 
 
 2.0.9 January 29, 2019